requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.0
murf==2.0.2
assemblyai
websockets==12.0
google-generativeai>=0.3.0
tavily-python>=0.3.0
requests>=2.25.0
pillow>=9.0.0

